Move config package to pkg
