+++ internal/server/options.go
@@ -3,7 +3,7 @@ package server
 import (
 	"net/url"
 
-	"github.com/kanopy-platform/grafana-auth-proxy/internal/config"
+	"github.com/kanopy-platform/grafana-auth-proxy/pkg/config"
 	"github.com/kanopy-platform/grafana-auth-proxy/pkg/grafana"
 )
