+++ pkg/grafana/client_test.go
@@ -5,7 +5,7 @@ import (
 
 	gapi "github.com/grafana/grafana-api-golang-client"
 	"github.com/grafana/grafana/pkg/models"
-	"github.com/kanopy-platform/grafana-auth-proxy/internal/config"
+	"github.com/kanopy-platform/grafana-auth-proxy/pkg/config"
 	"github.com/stretchr/testify/assert"
 )
